# app/public/docs/api.yaml
openapi: "3.0.3"
info:
  title: MagicWebb API
  version: "1.0.0"
  description: |
    MagicWebb is a decentralized NFT marketplace on Flare Coston2.
    All `/api/v1/*` routes are rate-limited (60 req/min per IP).
    Routes marked 🔒 require a valid SIWE JWT (Bearer token or `mw_s_*` cookie).
servers:
  - url: https://magicwebb.fly.dev
    description: Production (Flare Coston2 testnet)
  - url: http://localhost:8080
    description: Local development
paths:
...
  /api/v1/media:
    get:
      summary: Media proxy
      description: |
        Proxies external NFT images through same-origin with SSRF guards.
        Accepts http/https URLs (ipfs:// URIs are resolved server-side).
        When the URL is a local blob path, short-circuits to
        /api/v1/img/:sha256. All fetched images are self-hosted into the
        local Postgres BYTEA store on first access — no IPFS/Pinata needed.
      tags: [Media]
      parameters:
        - name: url
          in: query
          required: true
          schema: { type: string, example: "https://example.com/nft.png" }
      responses:
        "200":
          description: Image bytes with correct Content-Type
          headers:
            Cache-Control: { schema: { type: string, example: "public, max-age=86400" } }
        "400":
          description: Invalid or disallowed URL
          content: { text/plain: { schema: { type: string } } }
        "502":
          description: Upstream unavailable
...
  /api/v1/img/retry:
    post:
      summary: Trigger image self-hosting retry
      description: |
        Forces immediate fetch and self-host of a token's image from its
        upstream URL. Idempotent — returns "already_local" if already cached.
        Not JWT-protected; rate-limited by the api group (60 req/min/IP).
      tags: [Media]
      parameters:
        - name: coll
          in: query
          required: true
          schema: { type: string, example: "0x..." }
          description: Collection address
        - name: id
          in: query
          required: true
          schema: { type: string, example: "1" }
          description: Token ID
      responses:
        "200":
          description: Retry result
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [ok, already_local]
                  image_uri:
                    type: string
                    example: "/api/v1/img/abc123def..."
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          description: Token metadata not found or no image_uri
        "502":
          description: Upstream or self-host failed

  # ── Wallet ──────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: "SIWE JWT via Authorization: Bearer <token>"
    CookieAuth:
      type: apiKey
      in: cookie
      name: mw_s_
      description: "HttpOnly session cookie with addr prefix (e.g. mw_s_0xabcd...)"
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
...